requirements.txt
aiosqlite==0.19.0
discord.py==2.3.2
